Refactor layer functions in CausalSelfAttention, FeedForward, Transformer, and GPT modules to remove unnecessary name parameters

+++ lib/nano_ai.ex
@@ -24,6 +24,6 @@ defmodule NanoAi do
   end
 
   def explore(%Axon{} = model, template \\ {8, 1024}) do
-    Axon.Display.as_table(model, Nx.template(template, :bf16))
+    Axon.Display.as_table(model, Nx.template(template, :u32))
   end
 end
